SLA counter is now accurate
Since its triggered after the pipeline is complete
